requirements.txt
requests
fastapi
uvicorn
pydantic
mcp[cli]
langchain
langchain_community
langgraph 
langsmith
langchain_groq
pinecone
google-generativeai
python-dotenv